tests: organize tests
